# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Community & School Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed

  # Channels
...
  /admin/channels:
    post:
      operationId: createChannel
      tags: [channels, admin]
      summary: Create a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelRequest"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
  /admin/channels/{channelId}:
    patch:
      operationId: updateChannel
      tags: [channels, admin]
      summary: Update a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequest"
      responses:
        "200":
          description: Channel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      operationId: deleteChannel
      tags: [channels, admin]
      summary: Delete a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Channel deleted
  /admin/channels/{channelId}/reorder:
    post:
      operationId: reorderChannel
      tags: [channels, admin]
      summary: Reorder a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "200":
          description: Reordered

  # Posts
  /channels/{channelId}/posts:
    get:
      operationId: listPostsByChannel
      tags: [posts]
      summary: List posts in a channel
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /posts/{postId}/pin:
    post:
      operationId: pinPost
      tags: [posts]
      summary: Pin a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pinned
  /posts/{postId}/unpin:
    post:
      operationId: unpinPost
      tags: [posts]
      summary: Unpin a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unpinned

  # Comments
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create a comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # School - Browse
...
  /progress/lessons/{lessonId}:
    post:
      operationId: toggleLessonCompletion
      tags: [progress]
      summary: Toggle lesson completion
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleCompletionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonCompletion"
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created with temp password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
  /admin/users/{userId}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [admin]
      summary: Reset user password
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordResponse"
  /admin/users/{userId}/deactivate:
    patch:
      operationId: deactivateUser
      tags: [admin]
      summary: Deactivate a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deactivated

  # Uploads
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
        isActive:
          type: boolean
        forcePasswordChange:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - isActive
        - forcePasswordChange
        - createdAt
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
      required:
        - email
        - name
        - role
    CreateUserResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        tempPassword:
          type: string
      required:
        - user
        - tempPassword
    ResetPasswordResponse:
      type: object
      properties:
        tempPassword:
          type: string
      required:
        - tempPassword
    Channel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        adminsOnly:
          type: boolean
        isDefault:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - adminsOnly
        - isDefault
        - sortOrder
        - createdAt
    CreateChannelRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        adminsOnly:
          type: boolean
      required:
        - name
    UpdateChannelRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        adminsOnly:
          type: boolean
    ReorderRequest:
      type: object
      properties:
        direction:
          type: string
          enum: [up, down]
      required:
        - direction
    Post:
      type: object
      properties:
        id:
          type: integer
        channelId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        body:
          type: string
        isPinned:
          type: boolean
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - channelId
        - authorId
        - authorName
        - body
        - isPinned
        - commentCount
        - createdAt
    CreatePostRequest:
      type: object
      properties:
        body:
          type: string
      required:
        - body
    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorId
        - authorName
        - body
        - createdAt
    CreateCommentRequest:
      type: object
      properties:
        body:
          type: string
      required:
        - body
...
    LessonCompletion:
      type: object
      properties:
        lessonId:
          type: integer
        completed:
          type: boolean
      required:
        - lessonId
        - completed
    ToggleCompletionRequest:
      type: object
      properties:
        completed:
          type: boolean
      required:
        - completed
...